# Performance improvements to StatSerializer