--- a/lib/datadog/statsd/serialization/stat_serializer.rb
+++ b/lib/datadog/statsd/serialization/stat_serializer.rb
@@ -6,34 +6,24 @@ module Datadog
       class StatSerializer
         def initialize(prefix, global_tags: [])
           @prefix = prefix
+          @prefix_str = prefix.to_s
           @tag_serializer = TagSerializer.new(global_tags)
         end
 
         def format(name, delta, type, tags: [], sample_rate: 1)
-          String.new.tap do |stat|
-            stat << prefix if prefix
+          name = formated_name(name)
 
-            # stat value
-            stat << formated_name(name)
-            stat << ':'
-            stat << delta.to_s
-
-            # stat type
-            stat << '|'
-            stat << type
-
-            # sample_rate
-            if sample_rate != 1
-              stat << '|'
-              stat << '@'
-              stat << sample_rate.to_s
-            end
-
-            # tags
+          if sample_rate != 1
             if tags_list = tag_serializer.format(tags)
-              stat << '|'
-              stat << '#'
-              stat << tags_list
+              "#{@prefix_str}#{name}:#{delta}|#{type}|@#{sample_rate}|##{tags_list}"
+            else
+              "#{@prefix_str}#{name}:#{delta}|#{type}|@#{sample_rate}"
+            end
+          else
+            if tags_list = tag_serializer.format(tags)
+              "#{@prefix_str}#{name}:#{delta}|#{type}|##{tags_list}"
+            else
+              "#{@prefix_str}#{name}:#{delta}|#{type}"
             end
           end
         end
@@ -43,18 +33,21 @@ module Datadog
         end
 
         private
+
         attr_reader :prefix
         attr_reader :tag_serializer
 
         def formated_name(name)
-          formated = name.is_a?(String) ? name.dup : name.to_s
-
-          formated.tap do |f|
-            # replace Ruby module scoping with '.'
-            f.gsub!('::', '.')
-            # replace reserved chars (: | @) with underscores.
-            f.tr!(':|@', '_')
+          if name.is_a?(String)
+            # DEV: gsub is faster than dup.gsub!
+            formated = name.gsub('::', '.')
+          else
+            formated = name.to_s
+            formated.gsub!('::', '.')
           end
+
+          formated.tr!(':|@', '_')
+          formated
         end
       end
     end