@marcotc 2020-10-28
Another performance improvement for review when you have a chance, @kbogtob.